pyproject.toml
[project]
name = "cinema-monitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow>=12.0.0",
    "playwright>=1.56.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "httpx>=0.28.1",
    "beautifulsoup4>=4.12.3",
]

[project.scripts]
cinema-monitor = "src.main:main"

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "ruff>=0.6.0",
    "black>=24.10.0",
    "mypy>=1.11.2",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "B"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unused_configs = true

[tool.setuptools]
packages = ["src"]
